+++ classes/output/multi_user_course_report.php
@@ -26,6 +26,8 @@
 namespace report_embedquestion\output;
 use report_embedquestion\latest_attempt_table;
 use report_embedquestion\utils;
+use moodle_url;
+use stdClass;
 
 defined('MOODLE_INTERNAL') || die();
 
@@ -73,11 +75,26 @@ class multi_user_course_report {
     }
 
     /**
-     * Display the report.
+     * Display or download the report.
+     * @param string|null $download
      */
-    public function display_content() {
-        $table = new latest_attempt_table($this->context, $this->courseid, $this->groupid);
+    public function display_download_content($download = null) {
+        global $COURSE;
+        list ($filterform, $filter) = utils::get_filter_data(utils::get_url(['courseid' => $this->courseid]));
+        $filename = $COURSE->shortname . '_' . str_replace(' ', '_', $this->get_title());
+        if (!$download) {
+            $table = new latest_attempt_table($this->context, $this->courseid, $this->groupid, null, $filter);
+            // Display the filter form.
+            echo $filterform;
+            utils::allow_downloadability_for_attempt_table($table, $this->get_title(), $this->context);
+        } else {
+            $table = new latest_attempt_table($this->context, $this->courseid, $this->groupid, null, $filter, $download);
+            $table->is_downloading($download, $filename);
+            if ($table->is_downloading()) {
+                raise_memory_limit(MEMORY_EXTRA);
+            }
+        }
         $table->setup();
-        $table->out($this->pagesize, true, null);
+        $table->out($this->pagesize, true);
     }
 }