+++ changes.md
@@ -1,5 +1,11 @@
 # Change log for the Embedded questions progress report
 
+## Features in version 1.1
+
+* Ability to fileter the report by date. Either a from - to date range,
+  or last so many days or weeks.
+* Option to download the report in a range of formats.
+* Improvements to the report UI.
 
 ## Features in version 1.0